app/backend-flask/requirements.txt
Flask==3.1.0
gunicorn>=23.0.0
psycopg[binary,pool]>=3.2.5
flask-sqlalchemy>=3.1.1
Flask-Migrate>=4.1.0
structlog>=25.2.0
opentelemetry-distro>=0.51b0
opentelemetry-instrumentation-flask>=0.51b0
opentelemetry-exporter-otlp>=1.30.0
debugpy>=1.8.13
